refactor AccountablePeopleList into component
